+++ packages/SequenceTranslator/src/polytool/utils.ts
@@ -8,13 +8,13 @@ import {ALPHABET, ALIGNMENT, NOTATION} from '@datagrok-libraries/bio/src/utils/m
 
 export function _setPeptideColumn(col: DG.Column): void {
   addCommonTags(col);
-  col.setTag('units', NOTATION.SEPARATOR);
+  col.meta.units = NOTATION.SEPARATOR;
   col.setTag('separator', '-');
   // col.setTag('cell.renderer', 'sequence');
 }
 
 function addCommonTags(col: DG.Column<any>) {
-  col.setTag('quality', DG.SEMTYPE.MACROMOLECULE);
+  col.semType =DG.SEMTYPE.MACROMOLECULE;
   col.setTag('aligned', ALIGNMENT.SEQ);
   col.setTag('alphabet', ALPHABET.PT);
 }